Move command types to separate TypedCommand interface.
