+++ examples/value-state-machine/src/main/java/io/atomix/copycat/examples/DeleteCommand.java
@@ -15,14 +15,14 @@
  */
 package io.atomix.copycat.examples;
 
-import io.atomix.copycat.client.Command;
+import io.atomix.copycat.client.TypedCommand;
 
 /**
  * Value delete command.
  *
  * @author <a href="http://github.com/kuujo>Jordan Halterman</a>
  */
-public class DeleteCommand implements Command<Void> {
+public class DeleteCommand implements TypedCommand<Void> {
 
   @Override
   public Type type() {
