+++ test/src/test/java/io/atomix/copycat/test/ClusterTest.java
@@ -18,10 +18,7 @@ package io.atomix.copycat.test;
 import io.atomix.catalyst.transport.Address;
 import io.atomix.catalyst.transport.LocalServerRegistry;
 import io.atomix.catalyst.transport.LocalTransport;
-import io.atomix.copycat.client.Command;
-import io.atomix.copycat.client.ConnectionStrategies;
-import io.atomix.copycat.client.CopycatClient;
-import io.atomix.copycat.client.Query;
+import io.atomix.copycat.client.*;
 import io.atomix.copycat.client.session.Session;
 import io.atomix.copycat.server.Commit;
 import io.atomix.copycat.server.CopycatServer;
@@ -1243,7 +1240,7 @@ public class ClusterTest extends ConcurrentTestCase {
   /**
    * Test command.
    */
-  public static class TestCommand implements Command<String> {
+  public static class TestCommand implements TypedCommand<String> {
     private String value;
     private ConsistencyLevel consistency;
 
@@ -1292,7 +1289,7 @@ public class ClusterTest extends ConcurrentTestCase {
   /**
    * Test event.
    */
-  public static class TestEvent implements Command<String> {
+  public static class TestEvent implements TypedCommand<String> {
     private String value;
     private boolean own;
     private ConsistencyLevel consistency;
